Fix Create Account

+++ Assets/Scripts/Game_Entities/Job.cs
@@ -6,7 +6,7 @@ public class Job : MonoBehaviour
 {
     public string id { get; set; }
     public string Job_card_name { get; set; }
-    public float Children_cost { get; set; }
+    public int Children_cost { get; set; }
     public string Image_url { get; set; }
     public DateTime Create_at { get; set; }
     public int Create_by { get; set; }
@@ -16,7 +16,7 @@ public class Job : MonoBehaviour
     public List<Game_accounts> Game_accounts { get; set; }
 
 
-    public Job(string _id, string name, float children_cost, List<Game_accounts> gameAccounts)
+    public Job(string _id, string name, int children_cost, List<Game_accounts> gameAccounts)
     {
         this.Game_accounts = gameAccounts;
         this.id = _id;
